+++ src/preferences/general.hpp
@@ -165,6 +165,8 @@ namespace preferences {
 	std::string moved_color();
 	void set_moved_color(const std::string& color_id);
 
+	std::string disengaged_color();
+	void set_disengaged_color(const std::string& color_id);
 
 	bool show_allied_orb();
 	void set_show_allied_orb(bool show_orb);
@@ -181,6 +183,8 @@ namespace preferences {
 	bool show_partial_orb();
 	void set_show_partial_orb(bool show_orb);
 
+	bool show_disengaged_orb();
+	void set_show_disengaged_orb(bool show_orb);
 
 	bool use_color_cursors();
 	void _set_color_cursors(bool value);